Added a test utility module and extracted findUserActivationCode from user-service test into its own function
