+++ integration/services/user.service.spec.ts
@@ -5,6 +5,7 @@ import pool = require('../../src/dbutils/connection-pool');
 import q = require('../../src/dbutils/query');
 import errors = require('../../src/errors/errors');
 import models = require('../../src/models/models');
+import utils = require('../utils');
 import uuid = require('uuid');
 import r = require('rethinkdb');
 
@@ -152,13 +153,8 @@ describe('UserService', () => {
     var user:models.User;
     var activationCode:string;
     function findUserActivationCode(_user:models.User) {
-      var findUserActivationCodeQuery = r.db('froyo')
-        .table('activation')
-        .coerceTo('array')
-        .filter({'userId': _user.id})
-        .nth(0);
-      return q.run(findUserActivationCodeQuery)()
-        .then((result) => { activationCode = result.id})
+      return utils.findUserActivationCode(_user)
+        .then((ac) => {activationCode = ac;})
         .return(_user);
     }
     createUser(rs(), rs(), rs(), em(), rs(), rs())()
